Fix a weird issue with other libraries pulling in secrets.py
